Changed MicrOrm to use IDictionary<string,object> instead of dynamic.  dynamic caused memory leaks.

+++ MicrOrm/MicrOrm.PostgreSql.Test/ExecuteScalarTest.cs
@@ -52,7 +52,7 @@ namespace MicrOrm.PostgreSql.Test
       {
         var result = db.ExecuteScalar("SELECT test.get_user_id(:p0)", "Bob");
 
-        Assert.True(result > 0);
+        Assert.True((int)result > 0);
       }
     }
